requirements.txt
python-telegram-bot==21.5
httpx==0.27.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
aiohttp==3.9.5
apscheduler==3.10.4
packaging==24.1
pdfplumber==0.11.0
groq==1.0.0
python-docx==1.1.2
pytest==8.4.2
pytest-asyncio==1.2.0
